fix: rename rule
